requirements.txt
Flask
picamera
plotly
pyserial
numpy
pandas
RPi.GPIO
